renamed update_currently_watching.py to anime_list.py
